+++ src/graphdb.py
@@ -2,7 +2,6 @@
 This file contains functions for interacting with GraphDB
 """
 import os
-
 import requests
 
 admin_password = os.environ.get("ADMIN_PASSWORD", '')
@@ -15,18 +14,20 @@ def setup_graphdb():
     :return:
     """
     # Check if db exists
-    resp = requests.get(f"{endpoint}/size")
+    resp = requests.get(f"{endpoint}/size", timeout=60)
     if resp.status_code != 200:
         # GraphDB repository not created yet -- create it
         headers = {
             'Content-Type': 'text/turtle',
         }
-        requests.put(
-            f"{endpoint}",
-            headers=headers,
-            data=open(f"/var/www/skosmos-repository.ttl", "rb"),
-            auth=('admin', admin_password),
-        )
+        with open("/var/www/skosmos-repository.ttl", "rb") as fp:
+            requests.put(
+                f"{endpoint}",
+                headers=headers,
+                data=fp,
+                auth=('admin', admin_password),
+                timeout=60
+            )
         print(f"CREATED GRAPHDB[{endpoint}] DB[skosmos.tdb]")
     else:
         print(f"EXISTS GRAPHDB [{endpoint}]]")
@@ -37,8 +38,11 @@ def get_loaded_vocabs():
     Get all loaded vocabularies from GraphDB
     :return:
     """
-    graphs_response = requests.get(f"{endpoint}/rdf-graphs",
-                                   headers={"Accept": "application/json"})
+    graphs_response = requests.get(
+        f"{endpoint}/rdf-graphs",
+        headers={"Accept": "application/json"},
+        timeout=60
+    )
     tmp = []
     if graphs_response.status_code == 200:
         body = graphs_response.json()
@@ -51,6 +55,11 @@ def get_loaded_vocabs():
 
 
 def get_type(extension):
+    """
+    Get the http mimetype based on the extension of a file.
+    :param extension:
+    :return:
+    """
     if extension in ["ttl", "turtle"]:
         return "text/turtle"
     if extension in ["trig"]:
@@ -77,6 +86,7 @@ def add_vocabulary(graph, graph_name, extension):
         headers=headers,
         auth=('admin', admin_password),
         params={'context': f"<{graph_name}>"},
+        timeout=60,
     )
     print(f"RESPONSE: {response.status_code}")
     if response.status_code != 200:
